chore: avoid rm keep file (#3)
chore: avoid removeing keep file

+++ tests/basic_commands.rs
@@ -6,8 +6,12 @@ fn cleanup() {
     let examples_dir = current_dir.join("examples/to");
     let paths = std::fs::read_dir(&examples_dir).expect(&format!("Failed to read directory {:?}", examples_dir));
 
-    for path in paths {
+    for path in paths.into_iter() {
         let path = path.unwrap().path();
+        // ignore .gitkeep file
+        if path.ends_with(".gitkeep") {
+            continue;
+        }
         std::fs::remove_file(&path).expect(&format!("Failed to remove file {:?}", path));
     }
 }
